In parser check for empty line added. Enum moved to FruitTransaction as inner class. SupplyOperationHandler nad ReturnOperationHandler replaced by IncreaseOperationHandler.
