--- report.csv
@@ -1,3 +1,0 @@
-type,fruit,quantity
-banana,152
-apple,90
