+++ src/main/java/core/basesyntax/hadler/impl/IncreaseOperationHandler.java
@@ -4,10 +4,10 @@ import core.basesyntax.dao.StorageDao;
 import core.basesyntax.hadler.OperationHandler;
 import core.basesyntax.model.FruitTransaction;
 
-public class ReturnOperationHandler implements OperationHandler {
+public class IncreaseOperationHandler implements OperationHandler {
     private final StorageDao storageDao;
 
-    public ReturnOperationHandler(StorageDao storageDao) {
+    public IncreaseOperationHandler(StorageDao storageDao) {
         this.storageDao = storageDao;
     }
 
@@ -15,6 +15,6 @@ public class ReturnOperationHandler implements OperationHandler {
     public void handle(FruitTransaction transaction) {
         int storageQuantity = storageDao.getQuantity(transaction.getFruit());
         storageDao.save(transaction.getFruit(),
-                storageQuantity + transaction.getQuantity());
+                storageQuantity - transaction.getQuantity());
     }
 }
