+++ src/main/java/core/basesyntax/model/FruitTransaction.java
@@ -1,6 +1,7 @@
 package core.basesyntax.model;
 
-import core.basesyntax.enums.Operation;
+import java.util.Arrays;
+import java.util.NoSuchElementException;
 
 public class FruitTransaction {
     private Operation operation;
@@ -23,5 +24,25 @@ public class FruitTransaction {
 
     public int getQuantity() {
         return quantity;
+    }
+
+    public enum Operation {
+        BALANCE("b"),
+        SUPPLY("s"),
+        PURCHASE("p"),
+        RETURN("r");
+
+        private String value;
+        Operation(String value) {
+            this.value = value;
+        }
+
+        public static Operation getType(String value) {
+            return Arrays.stream(Operation.values())
+                    .filter(operation -> operation.value.equals(value))
+                    .findFirst()
+                    .orElseThrow(() ->
+                            new NoSuchElementException("Invalid operation type: " + value));
+        }
     }
 }
