+++ src/main/java/core/basesyntax/strategy/OperationStrategy.java
@@ -1,7 +1,7 @@
 package core.basesyntax.strategy;
 
-import core.basesyntax.enums.Operation;
 import core.basesyntax.hadler.OperationHandler;
+import core.basesyntax.model.FruitTransaction.Operation;
 
 public interface OperationStrategy {
     OperationHandler get(Operation operation);
